Separate read/write steps of cfn-flip
Fixes https://github.com/awslabs/aws-cfn-template-flip/issues/91

The current process can lead to losing your template when specifying the
same input and output file, due to the write action interfering with the
read action.

This change performs the flip function, storing the result in memory,
then writing it. This allows users to use the same file as an
inputand output.